AMATERASU-24 Refactor Spark out of Amaterasu executor to it's own project
