--- executor/src/test/resources/SparkSql/csv/SparkSqlTestCsv.csv
@@ -1,4 +1,0 @@
-name,age
-sampath,22
-kirupa,30
-dev,19
