--- executor/src/test/resources/SparkSql/json/SparkSqlTestData.json
@@ -1,3 +1,0 @@
-{"name":"Sampath","age":22}
-{"name":"Kirupa", "age":30}
-{"name":"Dev", "age":19}]
